refactor(cli): extract shared module from apply.rs and destroy.rs
Pull progress UI, the CLI observer, retry helpers, state write-back logic,
and small effect-execution helpers out of `commands/apply.rs` and
`commands/destroy.rs` into a new `commands/shared/` module. The split
matches the move map in #2265: progress.rs, observer.rs, state_writeback.rs,
effect_execution.rs, retry.rs.

Mechanical move only — no logic changes, no public-API changes. Helpers
that previously lived in `apply.rs` and were referenced by `destroy.rs`,
`plan.rs`, `state.rs`, `tests.rs`, and `wiring.rs` are now imported from
`crate::commands::shared::*` (`pub(crate)` visibility, since this is a
binary crate with no external consumers).

Closes #2265
