+++ carina-cli/src/commands/destroy.rs
@@ -23,8 +23,12 @@ use carina_core::parser::ProviderContext;
 
 use super::validate_and_resolve_with_config;
 use crate::DetailLevel;
-use crate::commands::apply::{
-    RefreshProgress, apply_name_overrides, format_duration, refresh_multi_progress, spinner_style,
+use crate::commands::shared::progress::{
+    RefreshProgress, format_duration, refresh_multi_progress, spinner_style,
+};
+use crate::commands::shared::retry::{WaitResult, is_retryable_delete_error, wait_for_deletion};
+use crate::commands::shared::state_writeback::{
+    apply_destroy_to_state, apply_name_overrides, build_orphan_resource,
 };
 use crate::commands::state::map_lock_error;
 use crate::display::{format_destroy_plan, format_effect};
@@ -916,107 +920,10 @@ async fn run_destroy_locked(
     }
 }
 
-/// Apply destroy results to the state file: remove destroyed resources and
-/// clear any exports (since exports reference attributes of destroyed resources).
-fn apply_destroy_to_state(state: &mut carina_state::StateFile, destroyed_ids: &[ResourceId]) {
-    for id in destroyed_ids {
-        state.remove_resource(&id.provider, &id.resource_type, id.name_str());
-    }
-    state.exports.clear();
-}
-
-/// Build a minimal `Resource` for an orphaned resource from the state file.
-///
-/// This creates a Resource with attributes reconstructed from state data,
-/// including `_binding` and `_dependency_bindings` so that dependency ordering
-/// and tree display work correctly.
-fn build_orphan_resource(sf: &carina_state::StateFile, id: &ResourceId) -> Resource {
-    let rs = sf
-        .find_resource(&id.provider, &id.resource_type, id.name_str())
-        .expect("orphan must exist in state file");
-    let attributes: HashMap<String, Value> = rs
-        .attributes
-        .iter()
-        .filter_map(|(k, v)| carina_core::value::json_to_dsl_value(v).map(|val| (k.clone(), val)))
-        .collect();
-    Resource {
-        id: id.clone(),
-        attributes: carina_core::resource::Expr::wrap_map(attributes),
-        kind: carina_core::resource::ResourceKind::Real,
-        lifecycle: rs.lifecycle.clone(),
-        prefixes: rs.prefixes.clone(),
-        binding: rs.binding.clone(),
-        dependency_bindings: rs.dependency_bindings.clone(),
-        module_source: None,
-        quoted_string_attrs: std::collections::HashSet::new(),
-    }
-}
-
-/// Check if a delete error is retryable due to implicit dependency ordering.
-///
-/// Some AWS errors indicate that a resource cannot be deleted yet because
-/// another resource still depends on it, even though there is no explicit
-/// ResourceRef dependency. These errors are retryable: once the blocker is
-/// deleted, the retry will succeed.
-fn is_retryable_delete_error(e: &carina_core::provider::ProviderError) -> bool {
-    if e.is_timeout {
-        return false;
-    }
-    let msg = e.to_string();
-    let retryable_patterns = [
-        "DependencyViolation",
-        "has dependent object",
-        "has a dependent object",
-        "resource has dependencies",
-        "mapped public address",
-        "Failed to detach",
-        // CloudControl operation timeout — often caused by dependent resources
-        // still being deleted (e.g., NAT Gateway blocking VPCGatewayAttachment)
-        "Exceeded attempts to wait",
-    ];
-    retryable_patterns.iter().any(|p| msg.contains(p))
-}
-
-/// Result of waiting for a resource deletion to complete.
-#[derive(Debug, PartialEq)]
-enum WaitResult {
-    /// Resource confirmed deleted (`state.exists == false`).
-    Deleted,
-    /// A `provider.read()` call returned an error.
-    ReadError(String),
-    /// The resource still existed after all retry attempts.
-    TimedOut,
-}
-
-/// Poll `provider.read()` in a loop until the resource disappears or an error /
-/// timeout occurs.
-///
-/// * `max_attempts` – how many times to poll (each preceded by `poll_interval`).
-/// * `poll_interval` – sleep duration between polls.
-async fn wait_for_deletion(
-    provider: &dyn Provider,
-    id: &ResourceId,
-    identifier: &str,
-    max_attempts: usize,
-    poll_interval: std::time::Duration,
-) -> WaitResult {
-    for _ in 0..max_attempts {
-        tokio::time::sleep(poll_interval).await;
-        match provider.read(id, Some(identifier)).await {
-            Ok(state) if !state.exists => return WaitResult::Deleted,
-            Ok(_) => {
-                // Still exists, keep waiting
-            }
-            Err(e) => return WaitResult::ReadError(e.to_string()),
-        }
-    }
-    WaitResult::TimedOut
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use carina_core::provider::{BoxFuture, ProviderError, ProviderResult};
+    use carina_core::provider::{BoxFuture, Provider, ProviderError, ProviderResult};
     use carina_core::resource::LifecycleConfig;
     use std::sync::atomic::{AtomicUsize, Ordering};
